feat: restructure into installable PyPI package (#1)
Split single-file linkding_server.py into a linkding_mcp_server/ package:
config (Settings + singleton), client (LinkDingClient), models, tools
(create_mcp_server factory), server (main entry), and setup. Add
pyproject.toml with build metadata, console entry points
(linkding-mcp-server, linkding-mcp-setup), AGPL-3.0 license, and ruff/
pytest config. Package now imports without requiring env vars at import
time. Add pytest suite (config/client/tools) with httpx mocked via respx.

Closes #1
