+++ linkding_mcp_server/__init__.py
@@ -1,0 +1,7 @@
+"""LinkDing MCP Server — an MCP server for the LinkDing bookmark manager."""
+
+__version__ = "0.1.0"
+
+from linkding_mcp_server.server import main
+
+__all__ = ["main", "__version__"]
