+++ pyproject.toml
@@ -1,0 +1,58 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "linkding-mcp-server"
+version = "0.1.0"
+description = "An MCP server for interacting with the LinkDing bookmark manager"
+readme = "README.md"
+requires-python = ">=3.12"
+license = {text = "AGPL-3.0-or-later"}
+authors = [
+    {name = "Jon Price", email = "jon@jonprice.io"},
+]
+keywords = ["mcp", "linkding", "bookmarks", "llm", "claude"]
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.12",
+    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
+    "Topic :: Internet :: WWW/HTTP",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+]
+dependencies = [
+    "fastmcp>=3.3.1",
+    "httpx>=0.25.0",
+    "python-dotenv>=1.2.2",
+    "pydantic>=2.13.4",
+]
+
+[project.urls]
+Homepage = "https://github.com/jon-the-dev/linkding-mcp-server"
+Repository = "https://github.com/jon-the-dev/linkding-mcp-server"
+
+[project.scripts]
+linkding-mcp-server = "linkding_mcp_server.server:main"
+linkding-mcp-setup = "linkding_mcp_server.setup:main"
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=8",
+    "pytest-asyncio>=0.23",
+    "ruff>=0.6",
+    "build>=1.2",
+    "respx>=0.21",
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["linkding_mcp_server*"]
+
+[tool.ruff]
+line-length = 100
+select = ["E", "F", "I"]
+ignore = ["E501"]
+
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
+testpaths = ["tests"]
